Finished a working version of POST courses

+++ src/main/kotlin/no/stunor/origo/organiserapi/model/courses/CourseVariant.kt
@@ -2,8 +2,8 @@ package no.stunor.origo.organiserapi.model.courses
 
 data class CourseVariant (
     var name: String?,
-    var length: Int?,
-    var climb: Int?,
+    var length: Double?,
+    var climb: Double?,
     var controls: List<Leg> = listOf(),
     var classes: List<String> = listOf()
 )
